Add Inspections property to Car main entity. Rename Inspection namespace, change classes correspondly.
