+++ CarRental.BussinessLayer/Managers/InspectionService.cs
@@ -1,6 +1,6 @@
 ﻿using CarRental.BussinessLayer.Interfaces;
 using CarRental.Data.Enums;
-using CarRental.Data.Inspection;
+using CarRental.Data.Checkup;
 using CarRental.Data.Models;
 using CarRental.Data.Models.Automobile;
 using System;
