+++ CarRental.BussinessLayer/Managers/InspectionsManager.cs
@@ -1,6 +1,6 @@
 ﻿using CarRental.BussinessLayer.Interfaces;
 using CarRental.Data.Enums;
-using CarRental.Data.Inspection;
+using CarRental.Data.Checkup;
 
 namespace CarRental.BussinessLayer.Managers;
 public class InspectionsManager : IInspectionsManager
